+++ src/frontend/ui/recherche-avancee/FiltreActiviteSanitaire.tsx
@@ -1,12 +1,15 @@
+import { sendEvent } from "@socialgouv/matomo-next";
 import { Dispatch, KeyboardEvent, SetStateAction, useContext, useEffect, useState } from "react";
 
 import { ActiviteSanitaire } from "./model/ActiviteSanitaire";
 import { classificationActiviteTypes } from "./model/ClassificationTypes";
 import styles from "./RechercheAvanceeFormulaire.module.css";
 import { WordingFr } from "../../configuration/wording/WordingFr";
+import { RECHERCHE_AVANCEE, RECHERCHE_AVANCEE_FILTRE_ACTIVITE } from "../../utils/nomenclature-matomo";
 import { ComparaisonContext } from "../commun/contexts/ComparaisonContext";
 import { RechercheAvanceeContext } from "../commun/contexts/RechercheAvanceeContext";
 import "@gouvfr/dsfr/dist/component/tooltip/tooltip.css";
+
 
 type FiltresForComparaisonProps = Readonly<{
   isComparaison: boolean;
@@ -122,6 +125,10 @@ export const FiltreActiviteSanitaire = ({ isComparaison, setIsChanged }: Filtres
   }
 
   const appliquerButton = () => {
+    if(!isComparaison){
+    sendEvent(RECHERCHE_AVANCEE);
+    sendEvent(RECHERCHE_AVANCEE_FILTRE_ACTIVITE);
+    }
     if (rechercheAvanceeContext) {
       rechercheAvanceeContext?.setActiviteMco(activiteMco.ranges);
       rechercheAvanceeContext?.setActivitePsy(activitePsy.ranges);