+++ CHANGELOG.md
@@ -24,6 +24,7 @@ hel-1050: Ajout d’event Matomo lors de l’export
 
 ### Features
 - hel-1059: [Compte doublon] Mise en place d'un cron pour vérifier la création
+
 - hel-1023: [Paramétrage]Historique des messages
 - hel-1022: [Paramétrage] Mise à jour du message d'accueil
 - hel-1026: [Paramétrage]Affichage du message
@@ -31,7 +32,6 @@ hel-1050: Ajout d’event Matomo lors de l’export
 - hel-1044: Ajouter la donnée etablissement principale/secondaire dans l’export des EJ
 - hel-1036: [RGPD] exercer les droits en matière de données personnelles
 - hel-1039: Ajout de Matomo
-- hel-1050: Ajout d’event Matomo lors de l’export
 
 ## [1.8.8](https://github.com/DNUM-SocialGouv/helios/compare/release/1.8.7...release/1.8.8) (2026-05-20)