+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Features
 hel-1050: Ajout d’event Matomo lors de l’export
+hel-1057: Ajout d'event Matomo lors de la consultation de la console admin
 
 ## [1.8.10](https://github.com/DNUM-SocialGouv/helios/compare/release/1.8.9...release/1.8.10) (2026-06-15)