+++ CHANGELOG.md
@@ -9,8 +9,12 @@
 - hel-1057: Ajout d'event Matomo lors de la consultation de la console admin
 - hel-1054: Ajout d'event Matomo lors de la consultation de la cartographie
 - hel-1053: Ajout d'event Matomo lors de la recherche simple
+<<<<<<< hel-1056/matomo-indicateur-acces-aide
 - hel-1056: Ajout d'event Matomo lors de la consultation de la page aide
 
+=======
+- hel-1048: Ajout d'event Matomo lors de la création d'une liste
+>>>>>>> feature/Matomo
 
 ## [1.8.10](https://github.com/DNUM-SocialGouv/helios/compare/release/1.8.9...release/1.8.10) (2026-06-15)